Update the crunch job for change from int to long.

+++ src/main/java/org/kitesdk/examples/movies/AnalyzeRatings.java
@@ -88,7 +88,7 @@ public class AnalyzeRatings extends CrunchTool implements Serializable {
   public static class GetRating extends MapFn<Rating, Integer> {
     @Override
     public Integer map(Rating rating) {
-      return rating.getRating();
+      return rating.getRating().intValue();
     }
   }
